[ AgdaFile ] new file and pretty AgdaFile.
